+++ src/main/java/com/kh/byulmee/mypage/controller/MyPageContoller.java
@@ -1,0 +1,82 @@
+package com.kh.byulmee.mypage.controller;
+
+import javax.servlet.http.HttpServletRequest;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.SessionAttributes;
+
+import com.kh.byulmee.member.model.vo.Member;
+
+@SessionAttributes("loginUser")
+
+@Controller
+public class MyPageContoller {
+	
+	@Autowired
+	private BCryptPasswordEncoder bcrypt;
+	
+	@RequestMapping("myPurView.me")
+	public String myPurView() {
+		return "myPur";
+	}
+	
+	@RequestMapping("myPurActView.me")
+	public String myPurActView() {
+		return "myPurAct";
+	}
+	
+	@RequestMapping("myPurProView.me")
+	public String myPurProView() {
+		return "myPurPro";
+	}
+	
+	@RequestMapping("myFavView.me")
+	public String myFavView() {
+		return "myFav";
+	}
+	
+	@RequestMapping("myFavActView.me")
+	public String myFavActView() {
+		return "myFavAct";
+	}
+	
+	@RequestMapping("myFavProView.me")
+	public String myFavProView() {
+		return "myFavPro";
+	}
+	
+	@RequestMapping("myFavStarView.me")
+	public String myFavStarView() {
+		return "myFavStar";
+	}
+	
+	@RequestMapping("myInfoView.me")
+	public String myInfoView() {
+		return "myInfoPwdCheck";
+	}
+	
+	@RequestMapping("myInfoPwdCheck.me")
+	public String myInfoPwdCheck(@RequestParam("memPwd")String memPwd, Model model, HttpServletRequest request) {
+
+		Member member = (Member) request.getSession().getAttribute("loginUser");
+
+		if(bcrypt.matches(memPwd, member.getMemPwd())) {
+			model.addAttribute("loginUser", member);
+			return "myInfo";
+		} else {
+			model.addAttribute("msg", "비밀번호가 일치하지 않습니다.");
+			model.addAttribute("url", "myInfoView.me");
+			return "../common/alert";
+		}
+	}
+	
+	@RequestMapping("myInfoUpdateView.me")
+	public String myInfoUpdateView() {
+		return "myInfoUpdate";
+	}
+}
